// src/services/openapi/processor-service/api.ts
export interface ProcessorPublishMessage {
    /**
     * 
     * @type {string}
     * @memberof ProcessorPublishMessage
     */
    topicName?: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessorPublishMessage
     */
    module?: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessorPublishMessage
     */
    process?: string;
    /**
     * 
     * @type {string}
     * @memberof ProcessorPublishMessage
     */
    eventName?: string;
    /**
     * 
     * @type {object}
     * @memberof ProcessorPublishMessage
     */
    data?: object;
}
/**
 * 
 * @export
 * @interface RequestByProcessIdDtoString
 */
export interface RequestByProcessIdDtoString {
    /**
     * 
     * @type {string}
     * @memberof RequestByProcessIdDtoString
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {string}
     * @memberof RequestByProcessIdDtoString
     */
    module?: string;
    /**
     * 
     * @type {string}
     * @memberof RequestByProcessIdDtoString
     */
    process?: string;
}
/**
 * 
 * @export
 * @interface SubmitRequestDto
 */
export interface SubmitRequestDto {
    /**
     * 
     * @type {string}
     * @memberof SubmitRequestDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {string}
     * @memberof SubmitRequestDto
     */
    module?: string;
    /**
     * 
     * @type {string}
     * @memberof SubmitRequestDto
     */
    process?: string;
    /**
     * 
     * @type {string}
     * @memberof SubmitRequestDto
     */
    action?: string;
    /**
     * 
     * @type {string}
     * @memberof SubmitRequestDto
     */
    comment?: string;
    /**
     * 
     * @type {boolean}
     * @memberof SubmitRequestDto
     */
    isCompleteEditAskForInfo?: boolean;
}
// ...
export class ProcessorControllerApi extends BaseAPI {
    /**
     * 
     * @summary API to publish message from http
     * @param {ProcessorPublishMessage} processorPublishMessage 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProcessorControllerApi
     */
    public httpPublishMessage(processorPublishMessage: ProcessorPublishMessage, options?: any) {
        return ProcessorControllerApiFp(this.configuration).httpPublishMessage(processorPublishMessage, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to submit bucket
     * @param {RequestByProcessIdDtoString} requestByProcessIdDtoString 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProcessorControllerApi
     */
    public stepperBucket(requestByProcessIdDtoString: RequestByProcessIdDtoString, options?: any) {
        return ProcessorControllerApiFp(this.configuration).stepperBucket(requestByProcessIdDtoString, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to submit bucket
     * @param {SubmitRequestDto} submitRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProcessorControllerApi
     */
    public submitBucketProcess(submitRequestDto: SubmitRequestDto, options?: any) {
        return ProcessorControllerApiFp(this.configuration).submitBucketProcess(submitRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to submit new workflow
     * @param {SubmitRequestDto} submitRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProcessorControllerApi
     */
    public submitNewWorkflow(submitRequestDto: SubmitRequestDto, options?: any) {
        return ProcessorControllerApiFp(this.configuration).submitNewWorkflow(submitRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

}